+++ src/system/number/Float.h
@@ -6,18 +6,18 @@
 /* ========================================================================= */
 
 struct Float_t {
-    float value;
+    double value;
 };
 
 CREATE_INITIALIZERS(Float);
-extern Float new_Float(float value);
+extern Float new_Float(double value);
 
 /* ========================================================================= */
-extern Float Float_plus_Float(float left, float right);
+extern Float Float_plus_Float(double left, double right);
 /* ========================================================================= */
 
-extern Float wrap_float(float value);
-extern float unwrap_float(Optr integer);
+extern Float wrap_float(double value);
+extern double unwrap_float(Optr integer);
 
 /* ========================================================================= */
